small bug fixes

+++ src/main/dco/ProjectDCO.java
@@ -1,3 +1,3 @@
 package main.dco;
 
-public record ProjectDCO(String project_title, String description,Integer icon_color_id) {}
+public record ProjectDCO(String project_title, String description, Long icon_color_id) {}
