+++ src/main/dco/RecurringTaskDCO.java
@@ -1,3 +1,3 @@
 package main.dco;
 
-public record RecurringTaskDCO(Integer task_id, Integer week_day_id) {}
+public record RecurringTaskDCO(Long task_id, Long week_day_id) {}
