+++ src/main/dco/TaskDCO.java
@@ -5,8 +5,8 @@ import java.time.LocalDate;
 public record TaskDCO(
 	String task_title,
 	String description,
-	Integer status_id,
+	Long status_id,
 	Integer priority,
 	LocalDate due_date,
-	Integer project_id
+	Long project_id
 ) {}
