+++ src/main/entities/Project.java
@@ -1,3 +1,3 @@
 package main.entities;
 
-public record Project(Integer project_id, String project_title, String description, Integer list_order, Integer icon_color_id) {}
+public record Project(Long project_id, String project_title, String description, Integer list_order, Long icon_color_id) {}
