+++ src/main/entities/RecurringTask.java
@@ -1,3 +1,3 @@
 package main.entities;
 
-public record RecurringTask(Integer task_id, Integer week_day_id) {}
+public record RecurringTask(Long task_id, Long week_day_id) {}
