+++ src/main/entities/WeekDays.java
@@ -1,3 +1,3 @@
 package main.entities;
 
-public record WeekDays(Integer week_day_id, String day_name) {}
+public record WeekDays(Long week_day_id, String day_name) {}
